+++ src/shell/BarcodeToPostcodeAction.js
@@ -1,10 +1,10 @@
-
-class BarcodeToPostcodeAction{
-    constructor(){
+class BarcodeToPostcodeAction {
+    constructor() {
         this.name = 'barcodeToPostcode';
         this.help = '\n * 条形码转换邮编的状态\n  a - 输入条形码并转换.\n  q - 返回主界面.';
     }
-    doAction(cmd){
+
+    doAction(cmd) {
         switch (cmd) {
             case 'a':
                 return 'dealBarcodeToPostcode';
